// src/types/status-code.enum.ts
export enum StatusCodes {
  OK = 200,
  CREATED = 201,
  NO_CONTENT = 202,
  NOT_MODIFIED = 304,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  METHOD_NOT_ALLOWED = 405,
  CONFLICT = 409,
  INTERNAL_SERVER_ERROR = 500,
  SERVER_TEMPORARILY_UNAVAILABLE = 503,
}

export enum ErrorCodes {
  BAD_REQUEST = "BAD_REQUEST",
  UNAUTHORIZED = "UNAUTHORIZED",
  FORBIDDEN = "FORBIDDEN",
  NOT_FOUND = "NOT_FOUND",
  METHOD_NOT_ALLOWED = "METHOD_NOT_ALLOWED",
  CONFLICT = "CONFLICT",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  SERVER_TEMPORARILY_UNAVAILABLE = "SERVER_TEMPORARILY_UNAVAILABLE",
}